requirements.txt

pytorch==1.0.2
torchvision==0.2.2
numpy
pillow
matplotlib
